KAN-23 / 영화 정보에서 카테고리 기본값 history 제거
